+++ src/components/ShoppingCartProductsContainer/index.stories.js
