+++ src/components/ShoppingCartProductsContainer/index.styles.js
@@ -1,0 +1,16 @@
+import styled from "styled-components";
+
+export const ShoppingCartProductsContainer = styled.div`
+  flex-grow: 1;
+  margin-right: 50px;
+  display: flex;
+  flex-direction: column;
+`;
+
+export const ProductsTotalQuantity = styled.p`
+  font-size: 20px;
+  line-height: 33px;
+  letter-spacing: 0.5px;
+  padding-bottom: 16px;
+  border-bottom: 4px solid #ddd;
+`;
