+++ src/components/ShoppingCartProductsController/index.styles.js
@@ -1,12 +1,5 @@
 import styled from "styled-components";
 
-export const ShoppingCartProductsContainer = styled.div`
-  flex-grow: 1;
-  margin-right: 50px;
-  display: flex;
-  flex-direction: column;
-`;
-
 export const ProductsControlContainer = styled.div`
   display: flex;
   justify-content: space-between;
@@ -31,12 +24,4 @@ export const ProductsRemoveButton = styled.button`
   border-radius: 4px;
   padding: 12px 22px;
   line-height: 21.33px;
-`;
-
-export const ProductsTotalQuantity = styled.p`
-  font-size: 20px;
-  line-height: 33px;
-  letter-spacing: 0.5px;
-  padding-bottom: 16px;
-  border-bottom: 4px solid #ddd;
 `;
